# Conflicts:
#	src/Nethermind/Nethermind.Core/BlockInfo.cs
#	src/Nethermind/Nethermind.Merge.Plugin/Handlers/V1/ForkchoiceUpdatedV1Handler.cs
#	src/Nethermind/Nethermind.Network/P2P/Subprotocols/Snap/Messages/ByteCodesMessage.cs

+++ src/Nethermind/Nethermind.Consensus/Processing/BlockchainProcessor.cs
@@ -60,6 +60,7 @@ namespace Nethermind.Consensus.Processing
 
         private int _currentRecoveryQueueSize;
         private readonly CompositeBlockTracer _compositeBlockTracer = new();
+        private Stopwatch _stopwatch = new();
 
         /// <summary>
         /// 
@@ -237,9 +238,7 @@ namespace Nethermind.Consensus.Processing
 
         private void RunProcessingLoop()
         {
-            _stats.Start();
-            if (_logger.IsDebug)
-                _logger.Debug($"Starting block processor - {_blockQueue.Count} blocks waiting in the queue.");
+            if (_logger.IsDebug) _logger.Debug($"Starting block processor - {_blockQueue.Count} blocks waiting in the queue.");
 
             FireProcessingQueueEmpty();
 
@@ -253,6 +252,7 @@ namespace Nethermind.Consensus.Processing
                 Block block = blockRef.Block;
 
                 if (_logger.IsTrace) _logger.Trace($"Processing block {block.ToString(Block.Format.Short)}).");
+                _stats.Start();
 
                 Block processedBlock = Process(block, blockRef.ProcessingOptions, _compositeBlockTracer.GetTracer());
                 if (processedBlock == null)
@@ -313,9 +313,8 @@ namespace Nethermind.Consensus.Processing
             ProcessingBranch processingBranch = PrepareProcessingBranch(suggestedBlock, options);
             PrepareBlocksToProcess(suggestedBlock, options, processingBranch);
 
-            Stopwatch stopwatch = Stopwatch.StartNew();
+            _stopwatch.Restart();
             Block[]? processedBlocks = ProcessBranch(processingBranch, suggestedBlock, options, tracer);
-            stopwatch.Stop();
             if (processedBlocks == null)
             {
                 return null;
@@ -343,7 +342,7 @@ namespace Nethermind.Consensus.Processing
                         $"Updating main chain: {lastProcessed}, blocks count: {processedBlocks.Length}");
                 _blockTree.UpdateMainChain(processingBranch.Blocks.ToArray(), true);
 
-                Metrics.LastBlockProcessingTimeInMs = stopwatch.ElapsedMilliseconds;
+                Metrics.LastBlockProcessingTimeInMs = _stopwatch.ElapsedMilliseconds;
             }
 
             if ((options & ProcessingOptions.MarkAsProcessed) != 0)
@@ -353,7 +352,7 @@ namespace Nethermind.Consensus.Processing
                         $"Marked blocks as processed {lastProcessed}, blocks count: {processedBlocks.Length}");
                 _blockTree.MarkChainAsProcessed(processingBranch.Blocks.ToArray());
 
-                Metrics.LastBlockProcessingTimeInMs = stopwatch.ElapsedMilliseconds;
+                Metrics.LastBlockProcessingTimeInMs = _stopwatch.ElapsedMilliseconds;
             }
 
             if ((options & ProcessingOptions.ReadOnlyChain) == ProcessingOptions.None)
@@ -455,7 +454,7 @@ namespace Nethermind.Consensus.Processing
 
             finally
             {
-                if (invalidBlockHash is not null)
+                if (invalidBlockHash is not null && (options & ProcessingOptions.ReadOnlyChain) == 0)
                 {
                     DeleteInvalidBlocks(invalidBlockHash);
                 }
